+++ app/src/main/AndroidManifest.xml
@@ -25,9 +25,9 @@
         </activity>
 
         <activity
-            android:name="com.example.starwars.screens.film.FilmDetailsActivity"
+            android:name="com.example.starwars.feature.film.presentation.FilmDetailsActivity"
             android:theme="@style/Theme.StarWars.NoActionBar"/>
-        <activity android:name="com.example.starwars.screens.film.FilmsActivity"/>
+        <activity android:name="com.example.starwars.feature.film.presentation.FilmsActivity"/>
 
         <activity android:name="feature.character.presentation.CharactersActivity"
             android:theme="@style/Theme.StarWars.NoActionBar">
