+++ app/src/main/res/layout/activity_films.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/img_bg"
-    tools:context=".screens.film.FilmsActivity"
+    tools:context="feature.film.presentation.FilmsActivity"
     android:fitsSystemWindows="true">
 
     <androidx.recyclerview.widget.RecyclerView
